refactor: consolidate signal-aware sleep helper
