+++ src/agents/embedded-agent-runner/run/attempt.async-tasks.ts
@@ -11,6 +11,7 @@ import {
   findTaskByRunIdForStatus,
   listTasksForOwnerOrRequesterSessionKeyForStatus,
 } from "../../../tasks/task-status-access.js";
+import { sleep } from "../../../utils/sleep.js";
 
 export type AsyncStartedToolMeta = {
   toolName?: string;
@@ -35,12 +36,6 @@ const COMPLETION_REQUIRED_TASK_KINDS = new Set([
 
 function resolveAsyncTaskPollIntervalMs(): number {
   return isFastTestRuntimeEnv() ? 10 : DEFAULT_ASYNC_TASK_POLL_INTERVAL_MS;
-}
-
-function sleep(ms: number): Promise<void> {
-  return new Promise((resolve) => {
-    setTimeout(resolve, Math.max(1, ms));
-  });
 }
 
 function createAbortError(signal: AbortSignal): Error {
@@ -194,7 +189,7 @@ export async function waitForCompletionRequiredAsyncTasks(params: {
   abortSignal?: AbortSignal;
 }): Promise<CompletionRequiredAsyncTaskWaitResult> {
   const now = params.now ?? Date.now;
-  const sleepFn = params.sleep ?? sleep;
+  const sleepFn = params.sleep ?? ((ms) => sleep(Math.max(1, ms)));
   const pollIntervalMs = params.pollIntervalMs ?? resolveAsyncTaskPollIntervalMs();
   const waitedRunIds = new Set<string>();
   const timedOutRunIds = new Set<string>();
